pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "mcp[cli]>=1.5.0",
    "openai>=1.69.0",
    "python-dotenv>=1.0.1",
]
